# openapi.yaml
openapi: 3.1.0
info:
  title: IFRS9 API
  version: 1.0.0
paths:
...
  /user/feedback:
    post:
      tags:
      - user actions
      summary: Create Feedback
      description: Create new feedback
      operationId: create_feedback_user_feedback_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - user actions
      summary: Get All Feedback
      description: Get all feedback visible to regular users
      operationId: get_all_feedback_user_feedback_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/FeedbackStatusEnum'
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeedbackResponse'
                title: Response Get All Feedback User Feedback Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /user/feedback/mine:
    get:
      tags:
      - user actions
      summary: Get My Feedback
      description: Get all feedback submitted by the currently logged in user
      operationId: get_my_feedback_user_feedback_mine_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/FeedbackResponse'
                type: array
                title: Response Get My Feedback User Feedback Mine Get
      security:
      - OAuth2PasswordBearer: []
  /user/help:
    post:
      tags:
      - user actions
      summary: Create Help
      description: Create new help request
      operationId: create_help_user_help_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HelpCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelpResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - user actions
      summary: Get My Help
      description: Get all help requests for the current user
      operationId: get_my_help_user_help_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/HelpStatusEnum'
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HelpResponse'
                title: Response Get My Help User Help Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /user/notifications:
    get:
      tags:
      - user actions
      summary: Get Notifications
      description: Get notifications for the current user
      operationId: get_notifications_user_notifications_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NotificationResponse'
                type: array
                title: Response Get Notifications User Notifications Get
      security:
      - OAuth2PasswordBearer: []
  /user/feedback/{feedback_id}:
    put:
      tags:
      - user actions
      summary: Update Feedback
      description: Update a specific feedback entry
      operationId: update_feedback_user_feedback__feedback_id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: feedback_id
        in: path
        required: true
        schema:
          type: integer
          title: Feedback Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '404':
          description: Feedback not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - user actions
      summary: Delete Feedback
      description: Delete a specific feedback entry
      operationId: delete_feedback_user_feedback__feedback_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: feedback_id
        in: path
        required: true
        schema:
          type: integer
          title: Feedback Id
      responses:
        '204':
          description: Successful Response
        '404':
          description: Feedback not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - user actions
      summary: Get Feedback
      description: Get a specific feedback entry by ID
      operationId: get_feedback_user_feedback__feedback_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: feedback_id
        in: path
        required: true
        schema:
          type: integer
          title: Feedback Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '404':
          description: Feedback not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /user/feedback/{feedback_id}/like:
    post:
      tags:
      - user actions
      summary: Like Feedback
      description: Like feedback or unlike if already liked
      operationId: like_feedback_user_feedback__feedback_id__like_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: feedback_id
        in: path
        required: true
        schema:
          type: integer
          title: Feedback Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '404':
          description: Feedback not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /user/help/{help_id}:
    put:
      tags:
      - user actions
      summary: Update Help
      description: Update a specific help request
      operationId: update_help_user_help__help_id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: help_id
        in: path
        required: true
        schema:
          type: integer
          title: Help Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HelpUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelpResponse'
        '404':
          description: Help not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - user actions
      summary: Delete Help
      description: Delete a help request
      operationId: delete_help_user_help__help_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: help_id
        in: path
        required: true
        schema:
          type: integer
          title: Help Id
      responses:
        '204':
          description: Successful Response
        '404':
          description: Help not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - user actions
      summary: Get Help
      description: Get a specific help request by ID
      operationId: get_help_user_help__help_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: help_id
        in: path
        required: true
        schema:
          type: integer
          title: Help Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelpResponse'
        '404':
          description: Help not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    FeedbackCreate:
      properties:
        description:
          type: string
          minLength: 3
          title: Description
      type: object
      required:
      - description
      title: FeedbackCreate
      description: Schema for creating new feedback
    FeedbackResponse:
      properties:
        description:
          type: string
          minLength: 3
          title: Description
        id:
          type: integer
          title: Id
        status:
          $ref: '#/components/schemas/FeedbackStatusEnum'
        user_id:
          type: integer
          title: User Id
        user:
          anyOf:
          - $ref: '#/components/schemas/UserBasic'
          - type: 'null'
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
        is_creator:
          type: boolean
          title: Is Creator
        like_count:
          type: integer
          title: Like Count
          default: 0
        is_liked_by_user:
          type: boolean
          title: Is Liked By User
          default: false
      type: object
      required:
      - description
      - id
      - status
      - user_id
      - created_at
      - is_creator
      title: FeedbackResponse
      description: Schema for feedback response
    FeedbackStatusEnum:
      type: string
      enum:
      - submitted
      - open
      - closed
      - returned
      - in development
      - completed
      title: FeedbackStatusEnum
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HelpCreate:
      properties:
        description:
          type: string
          minLength: 10
          title: Description
      type: object
      required:
      - description
      title: HelpCreate
...
    HelpStatusEnum:
      type: string
      enum:
      - submitted
      - open
      - closed
      - returned
      - in development
      - completed
      title: HelpStatusEnum
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: token